Simplify generation of wp-config.php.
Instead of a bunch of perl calls, just do a big echo and use
variable interpolation. This should be easier to read, and will be
idempotent if we extract it into a separate script.

Also, this change generates the Authentication Unique Keys and Salts
properly instead of leaving them unset.